# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: НейроКлип API — AI video production platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{id}/generate-script:
    post:
      operationId: generateScript
      tags: [pipeline]
      summary: Run AI script generation (mock)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}/generate-images:
    post:
      operationId: generateImages
      tags: [pipeline]
      summary: Run AI image generation (mock)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}/generate-audio:
    post:
      operationId: generateAudio
      tags: [pipeline]
      summary: Run TTS audio generation (mock)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}/render:
    post:
      operationId: renderVideo
      tags: [pipeline]
      summary: Run final render (mock — instantly produces video URL)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
  /projects/{id}/progress:
    get:
      operationId: getProjectProgress
      tags: [pipeline]
      summary: Get current render progress
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectProgress" }
...
components:
...
  schemas:
...
    Project:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        topicDescription: { type: string }
        category:
          type: string
          enum: [educational, historical, content, marketing, news, story, vlog, howto, kids, business]
        targetDurationSec: { type: integer }
        durationSec: { type: integer }
        visualStyle: { type: string }
        voiceId: { type: string }
        voiceSpeed: { type: number }
        backgroundMusicId: { type: ["string", "null"] }
        musicVolume: { type: integer }
        addSubtitles: { type: boolean }
        status: { type: string, enum: [draft, script_ready, images_ready, audio_ready, rendering, done, failed] }
        currentStep: { type: integer, minimum: 1, maximum: 6 }
        aspectRatio: { type: string, enum: ["16:9", "9:16", "1:1"] }
        shareToken: { type: ["string", "null"] }
        parentProjectId: { type: ["string", "null"] }
        finalVideoUrl: { type: ["string", "null"] }
        thumbnailUrl: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        scenes:
          type: array
          items: { $ref: "#/components/schemas/Scene" }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - title
        - topicDescription
        - targetDurationSec
        - durationSec
        - visualStyle
        - voiceId
        - voiceSpeed
        - musicVolume
        - addSubtitles
        - status
        - currentStep
        - scenes
        - category
        - createdAt
        - updatedAt
...
    Scene:
      type: object
      properties:
        id: { type: string }
        projectId: { type: string }
        orderIndex: { type: integer }
        title: { type: string }
        narration: { type: string }
        imagePrompt: { type: string }
        imageUrl: { type: ["string", "null"] }
        audioUrl: { type: ["string", "null"] }
        durationSec: { type: number }
        animationType: { type: string }
        transitionType: { type: string }
      required: [id, projectId, orderIndex, title, narration, imagePrompt, durationSec, animationType, transitionType]
    ProjectProgress:
      type: object
      properties:
        projectId: { type: string }
        status: { type: string }
        currentStep: { type: integer }
        progress: { type: integer, minimum: 0, maximum: 100 }
        stageLabel: { type: string }
        etaSeconds: { type: integer }
        queuePosition: { type: integer }
      required: [projectId, status, currentStep, progress, stageLabel, etaSeconds, queuePosition]
...